+++ src/containers/Layout/Header.tsx
@@ -46,7 +46,7 @@ const Header: React.FC<IHeaderProps> = ({ onNavToggle, user }: IHeaderProps) => 
         <Badge key={3} style={BadgeStyle}>
           <span>
             Latest update:{" "}
-            <Moment format="DD MMM YYYY @ HH:mm">{`2021-06-08T13:30:10.297464-04:00`}</Moment>
+            <Moment format="DD MMM YYYY @ HH:mm">{`2021-06-14T15:40:10.297464-04:00`}</Moment>
           </span>
         </Badge>
       </React.Fragment>
