+++ src/pages/WorkflowsPage/index.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { useDispatch } from "react-redux";
 import Wrapper from "../../containers/Layout/PageWrapper";
-import StudyList from "./StudyList";
+
 import FileDetails from "./FileDetails";
-import PatientLookup from "./PatientLookup";
+
 import { setSidebarActive } from "../../store/ui/actions";
 import { resetWorkflowState } from "../../store/workflows/actions";
 import "./Workflows.scss";
@@ -28,8 +28,6 @@ const WorkflowsPage = () => {
 
   return (
     <Wrapper>
-      <PatientLookup />
-      <StudyList />
       <FileDetails />
     </Wrapper>
   );
